requirements2.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.7.4.post0
altair==5.5.0
annotated-types==0.7.0
anyio==4.3.0
app==0.0.1
app-config==1.0.1
argcomplete==1.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
aspose-words==24.12.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==3.0.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.8.0
bleach==6.1.0
blinker==1.6.3
boto==2.49.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2021.10.8
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.1.7
cobble==0.1.4
colorama==0.4.6
comm==0.2.1
contourpy==1.1.0
cssselect2==0.7.0
cycler==0.11.0
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
deepcut==0.7.0.0
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djongo==1.3.7
dm-tree==0.1.8
docx2pdf==0.1.8
docx2txt==0.8
EbookLib==0.17.1
exceptiongroup==1.2.0
executing==2.0.1
extract-msg==0.23.1
fastapi==0.115.5
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.12.2
fire==0.7.0
Flask==3.0.0
flatbuffers==24.3.7
fonttools==4.41.0
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.2
IMAPClient==2.1.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.29.3
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.13
keras==3.0.5
kiwisolver==1.4.4
libclang==16.0.6
lxml==5.3.0
mammoth==1.9.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mongoengine==0.29.1
mpmath==1.3.0
multidict==5.1.0
namex==0.0.7
narwhals==1.15.1
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nbimporter==0.3.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.25.1
olefile==0.46
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.12
overrides==7.7.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pdf2docx==0.5.8
pdfminer.six==20181108
Pillow==10.0.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.1.0
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
Pygments==2.17.2
pymongo==3.12.3
PyMuPDF==1.25.1
pypandoc==1.14
pyparsing==3.0.9
pypdf==5.1.0
pyphen==0.17.0
pyspell==0.0.6
pyspellchecker==0.8.1
pythainlp==5.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==0.19.0
python-json-logger==2.0.7
python-multipart==0.0.19
python-pptx==0.6.18
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
quanto==0.2.0
referencing==0.33.0
regex==2024.4.16
reportlab==4.2.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.8.1
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.12.0
sklearn==0.0.post7
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.2
sympy==1.12
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.0
textract==1.6.3
threadpoolctl==3.2.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.40.0
typer==0.14.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==1.5.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
wcwidth==0.2.13
weasyprint==63.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.6.3
zipp==3.16.2
zope.interface==7.0.1
zopfli==0.2.3.post1
